metricas_estreaming/requirements.txt
requests
scapy
matplotlib
numpy
selenium
beautifulsoup4
pymongo
python-dotenv
